+++ packages/gov-ui-kit/CHANGELOG.md
@@ -16,6 +16,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 -   Remove fixed width from `EmptyState` core component
 -   Center `CardEmptyState` core component
 -   Truncate long strings on `DaoDataListItem`, `AssetDataListItem`, `VoteDataListItem` module components
+-   Fix errors and warnings thrown on component tests
 
 ### Changed